Fix clippy warnings in examples (32→21 remaining)
Fixed via cargo clippy --fix and manual fixes:
- clone() -> std::slice::from_ref() for single-element slices
- empty println!("") -> println!()
- removed unused std::io::Write imports
- removed unused std::ptr imports
- clamp-like patterns auto-fixed
- unnecessary casts removed

Remaining 21 are cosmetic examples-level only:
- cfg v4l2: platform-specific (expected)
- Vec::new()+push: vec![] macro suggestion
- complex types: type alias suggestions
- identical if blocks: dead code paths

Library code: 0 warnings. Tests: 540 pass.

Total: 540 lib + 24 integration + 11 property = 575 tests, 0 warnings

+++ cam-rust/cam-isp/examples/full_pipeline.rs
@@ -41,7 +41,7 @@ fn main() {
         .with_blend_weight(0.6);
 
     // 7. Display: RGB float output
-    let display = DisplayBlock::new(w as u32);
+    let display = DisplayBlock::new(w);
 
     // Compose ONNX
     let blocks: Vec<&dyn cam_isp::pipeline::IspBlock> = vec![
